+++ SpatialGDK/Source/SpatialGDKServices/Private/Interop/Connection/EditorWorkerController.cpp
@@ -1,16 +1,13 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
-#pragma once
+#include "Interop/Connection/EditorWorkerController.h"
 
-#if WITH_EDITOR
+#include "SpatialGDKServicesPrivate.h"
 
 #include "Editor.h"
-#include "Modules/ModuleManager.h"
-#include "Settings/LevelEditorPlaySettings.h"
-#include "SpatialGDKEditorToolbar.h"
 
-namespace SpatialGDK
+#if WITH_EDITOR
+namespace
 {
-
 struct EditorWorkerController
 {
 	void OnPrePIEEnded(bool bValue)
@@ -22,15 +19,11 @@ struct EditorWorkerController
 	void OnSpatialShutdown()
 	{
 		LastPIEEndTime = 0;	// Reset PIE end time to ensure replace-a-worker isn't called
-		FSpatialGDKEditorToolbarModule& Toolbar = FModuleManager::GetModuleChecked<FSpatialGDKEditorToolbarModule>("SpatialGDKEditorToolbar");
-		Toolbar.OnSpatialShutdown.Remove(SpatialShutdownHandle);
 		FEditorDelegates::PrePIEEnded.Remove(PIEEndHandle);
 	}
 
-	void InitWorkers(const FString& WorkerType)
+	void InitWorkers()
 	{
-		ReplaceProcesses.Empty();
-
 		// Only issue the worker replace request if there's a chance the load balancer hasn't acknowledged
 		// that the previous session's workers have disconnected. There's no hard `heartbeat` time for this as
 		// it's dependent on multiple factors (fabric load etc.), so this value was landed on after significant
@@ -38,27 +31,30 @@ struct EditorWorkerController
 		const int64 WorkerReplaceThresholdSeconds = 8;
 
 		int64 SecondsSinceLastSession = FDateTime::Now().ToUnixTimestamp() - LastPIEEndTime;
-		UE_LOG(LogSpatialWorkerConnection, Verbose, TEXT("Seconds since last session - %d"), SecondsSinceLastSession);
+		UE_LOG(LogSpatialGDKServices, Verbose, TEXT("Seconds since last session - %d"), SecondsSinceLastSession);
 
 		PIEEndHandle = FEditorDelegates::PrePIEEnded.AddRaw(this, &EditorWorkerController::OnPrePIEEnded);
 
-		FSpatialGDKEditorToolbarModule& Toolbar = FModuleManager::GetModuleChecked<FSpatialGDKEditorToolbarModule>("SpatialGDKEditorToolbar");
-		SpatialShutdownHandle = Toolbar.OnSpatialShutdown.AddRaw(this, &EditorWorkerController::OnSpatialShutdown);
+		const ULevelEditorPlaySettings* LevelEditorPlaySettings = GetDefault<ULevelEditorPlaySettings>();
+		const int32 WorkerCount = LevelEditorPlaySettings->GetTotalPIEServerWorkerCount();
+		WorkerIds.SetNum(WorkerCount);
+		ReplaceProcesses.SetNum(WorkerCount);
 
-		int32 PlayNumberOfServers;
-		GetDefault<ULevelEditorPlaySettings>()->GetPlayNumberOfServers(PlayNumberOfServers);
-
-		WorkerIds.SetNum(PlayNumberOfServers);
-		for (int i = 0; i < PlayNumberOfServers; ++i)
+		int32 WorkerIdIndex = 0;
+		for (const TPair<FName, int32>& WorkerType : LevelEditorPlaySettings->GetPIEServerWorkers())
 		{
-			FString NewWorkerId = WorkerType + FGuid::NewGuid().ToString();
-
-			if (!WorkerIds[i].IsEmpty() && SecondsSinceLastSession < WorkerReplaceThresholdSeconds)
+			for (int i = 0; i < WorkerType.Value; ++i)
 			{
-				ReplaceProcesses.Add(ReplaceWorker(WorkerIds[i], NewWorkerId));
-			}
+				const FString NewWorkerId = WorkerType.Key.ToString() + FGuid::NewGuid().ToString();
 
-			WorkerIds[i] = NewWorkerId;
+				if (!WorkerIds[WorkerIdIndex].IsEmpty() && SecondsSinceLastSession < WorkerReplaceThresholdSeconds)
+				{
+					ReplaceProcesses.Add(ReplaceWorker(WorkerIds[WorkerIdIndex], NewWorkerId));
+				}
+
+				WorkerIds[WorkerIdIndex] = NewWorkerId;
+				WorkerIdIndex++;
+			}
 		}
 	}
 
@@ -96,6 +92,31 @@ struct EditorWorkerController
 	FDelegateHandle SpatialShutdownHandle;
 };
 
-} // namespace SpatialGDK
+static EditorWorkerController WorkerController;
+} // end namespace
 
-#endif
+namespace SpatialGDKServices
+{
+void InitWorkers(bool bConnectAsClient, FString& OutWorkerId)
+{
+	const bool bSingleThreadedServer = !bConnectAsClient && (GPlayInEditorID > 0);
+	const int32 FirstServerEditorID = 1;
+	if (bSingleThreadedServer)
+	{
+		if (GPlayInEditorID == FirstServerEditorID)
+		{
+			WorkerController.InitWorkers();
+		}
+
+		WorkerController.BlockUntilWorkerReady(GPlayInEditorID - 1);
+		OutWorkerId = WorkerController.WorkerIds[GPlayInEditorID - 1];
+	}
+}
+
+void OnSpatialShutdown()
+{
+	WorkerController.OnSpatialShutdown();
+}
+
+} // namespace SpatialGDKServices
+#endif // WITH_EDITOR