+++ SpatialGDK/Source/SpatialGDKServices/SpatialGDKServices.Build.cs
@@ -1,0 +1,20 @@
+// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
+
+using UnrealBuildTool;
+
+public class SpatialGDKServices : ModuleRules
+{
+	public SpatialGDKServices(ReadOnlyTargetRules Target) : base(Target)
+	{
+		PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+		bFasterWithoutUnity = true;
+
+		PrivateDependencyModuleNames.AddRange(
+			new string[] {
+				"Core",
+				"CoreUObject",
+				"EngineSettings",
+				"UnrealEd"
+			});
+	}
+}